pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tie"
version = "0.1.0"
description = "Lightweight internationalization library"
authors = [{name = "Maksim Panov", email = "panovo@essent.space"}]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = ["pyyaml", "packaging", "typing_extensions"]

